+++ Changelog.md
@@ -3,6 +3,7 @@
 ## Version 0.1.2
 
  - README updates.
+ - More browser support
  - Always run tests in build process
 
 ## Version 0.1.1