+++ Changelog.md
@@ -5,6 +5,7 @@
  - README updates.
  - More browser support
  - Always run tests in build process
+ - Add `uriCallback` method of getting `URI`
 
 ## Version 0.1.1