+++ Changelog.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Version 0.1.2
+
+ - README updates.
+ - More browser support
+ - Always run tests in build process
+ - Add `uriCallback` method of getting `URI`
+ - Passing full URLs to URI.query
+ - Getting Location-like objects by simply calling `URI`. 
+ - Add a search parameter to the Location-like object.
+
 ## Version 0.1.1
 
  - Documentation Updates